Move test resources out of test dir
