+++ chronograf/ui/src/flux/ast/stubs/from.ts
