+++ chronograf/ui/src/flux/ast/stubs/variable.ts
