+++ chronograf/ui/src/flux/ast/stubs/variables.ts
