+++ chronograf/ui/src/flux/ast/stubs/varsAndExpressions.ts
