+++ chronograf/ui/src/flux/helpers/stubs/suggestions.ts
