+++ chronograf/ui/src/shared/parsing/flux/constants.ts
