+++ chronograf/ui/src/tempVars/components/TemplateControlBar.test.tsx
@@ -5,7 +5,7 @@ import TemplateControlBar from 'src/tempVars/components/TemplateControlBar'
 import TemplateControl from 'src/tempVars/components/TemplateControl'
 import TemplateVariableEditor from 'src/tempVars/components/TemplateVariableEditor'
 import OverlayTechnology from 'src/reusable_ui/components/overlays/OverlayTechnology'
-import {source} from 'test/resources'
+import {source} from 'mocks/dummy'
 
 import {TemplateType, TemplateValueType} from 'src/types'
